Preserve the politician data set on kuorum platform instead CSV file

+++ grails-app/services/kuorum/users/PoliticianService.groovy
@@ -233,7 +233,7 @@ class PoliticianService {
             log.info("Updating ${politician.name} with externalId: ${externalId}")
         }else{
             politician = new KuorumUser()
-            politician.enabled = false
+            politician.enabled = politician.enabled?:false
             politician.setPoliticianExtraInfo(new PoliticianExtraInfo(externalId: externalId))
             log.info("Creating politician with ipdbId ${externalId}")
         }
@@ -242,15 +242,15 @@ class PoliticianService {
 
     private void populateBasicData(KuorumUser politician, def line){
         politician.name = line."name"
-        politician.bio = line."bio"
+        politician.bio = politician.bio?:line."bio"
         politician.userType = UserType.POLITICIAN
         politician.email = line."email"?:politician.email?:"info+${line.id}-${politician.name.encodeAsMD5()}@kuorum.org"
-        politician.password = "CSV ${Math.random()}"
+        politician.password = !politician.password?.startsWith("CSV")?"CSV ${Math.random()}":politician.password
         politician.verified = true
         politician.personalData = politician.personalData ?: new PersonData()
         politician.personalData.gender = line."gender"?Gender.valueOf(line."gender"):Gender.MALE
         politician.personalData.userType = politician.userType
-        politician.personalData.telephone = line."phone"
+        politician.personalData.telephone = politician?.personalData?.telephone?:line."phone"
         if (!politician.save()){
             throw new KuorumException("Basic data not porvided, ${politician.errors}")
         }
@@ -258,7 +258,7 @@ class PoliticianService {
 
     private void populateImages(KuorumUser politician, def line){
         String avatarUrl = line."picture"
-        if (avatarUrl.startsWith("http")){
+        if (!politician.avatar && avatarUrl.startsWith("http")){
             KuorumFile avatar = fileService.createExternalFile(politician, avatarUrl,FileGroup.USER_AVATAR, FileType.IMAGE)
             if (politician.avatar && avatar){
                 fileService.deleteKuorumFile(politician.avatar)
@@ -266,7 +266,7 @@ class PoliticianService {
             politician.avatar = avatar
         }
         String profileUrl = line."politicalPartyImage"
-        if (profileUrl.startsWith("http")){
+        if (!politician.imageProfile && profileUrl.startsWith("http")){
             KuorumFile imageProfile = fileService.createExternalFile(politician, profileUrl,FileGroup.USER_PROFILE, FileType.IMAGE)
             if (politician.imageProfile && imageProfile){
                 fileService.deleteKuorumFile(politician.imageProfile)
@@ -279,15 +279,15 @@ class PoliticianService {
         if (!politician.socialLinks){
             politician.socialLinks = new SocialLinks()
         }
-        politician.socialLinks.blog = line."blog"?.trim()
-        politician.socialLinks.facebook = line."facebook"?.trim()
-        politician.socialLinks.twitter = line."twitter"?.trim()
-        politician.socialLinks.linkedIn = line."linkedin"?.trim()
-        politician.socialLinks.googlePlus= line."googlePlus"?.trim()
-        politician.socialLinks.instagram = line."instagram"?.trim()
-        politician.socialLinks.youtube= line."youtubeChannel"?.trim()
-        politician.socialLinks.officialWebSite = line."officialWebsite"?.trim()
-        politician.socialLinks.institutionalWebSite = line."sourceWebsite"?.trim()
+        politician.socialLinks.blog = politician.socialLinks.blog?:line."blog"?.trim()
+        politician.socialLinks.facebook = politician.socialLinks.facebook?:line."facebook"?.trim()
+        politician.socialLinks.twitter = politician.socialLinks.twitter?:line."twitter"?.trim()
+        politician.socialLinks.linkedIn = politician.socialLinks.linkedIn?:line."linkedin"?.trim()
+        politician.socialLinks.googlePlus= politician.socialLinks.googlePlus?:line."googlePlus"?.trim()
+        politician.socialLinks.instagram = politician.socialLinks.instagram?:line."instagram"?.trim()
+        politician.socialLinks.youtube= politician.socialLinks.youtube?:line."youtubeChannel"?.trim()
+        politician.socialLinks.officialWebSite = politician.socialLinks.officialWebSite?:line."officialWebsite"?.trim()
+        politician.socialLinks.institutionalWebSite = politician.socialLinks.institutionalWebSite?:line."sourceWebsite"?.trim()
     }
 
     private void populateExternalPoliticianActivity(KuorumUser politician, def line){
@@ -345,21 +345,21 @@ class PoliticianService {
         if (!politician.politicianExtraInfo){
             throw RuntimeException("Este politico no tiene el id de la BBDD")
         }
-        politician.politicianExtraInfo.completeName = line."completeName"?.trim()?:line."name"?.trim()
-        politician.politicianExtraInfo.birthDate = parseDate(line."dateOfBirth", "dd/MM/yyyy")
-        politician.politicianExtraInfo.birthPlace = line."placeOfBirth"?.trim()
-        politician.politicianExtraInfo.family = line."family"?.trim()
+        politician.politicianExtraInfo.completeName =politician?.politicianExtraInfo?.completeName?: line."completeName"?.trim()?:line."name"?.trim()
+        politician.politicianExtraInfo.birthDate = politician.politicianExtraInfo.birthDate?:parseDate(line."dateOfBirth", "dd/MM/yyyy")
+        politician.politicianExtraInfo.birthPlace = politician.politicianExtraInfo.birthPlace?:line."placeOfBirth"?.trim()
+        politician.politicianExtraInfo.family = politician.politicianExtraInfo.family?:line."family"?.trim()
     }
 
     private void populateProfessionalDetails(KuorumUser politician, def line){
         if (!politician.professionalDetails){
             politician.professionalDetails = new ProfessionalDetails()
         }
-        politician.professionalDetails.politicalParty = line."politicalParty"
-        politician.professionalDetails.position = line."position"
-        politician.professionalDetails.institution = line."institution"
-        politician.professionalDetails.constituency = findConstituency(line)
-        politician.professionalDetails.region =findRegion(line)
+        politician.professionalDetails.politicalParty = politician.professionalDetails.politicalParty?:line."politicalParty"?.trim()
+        politician.professionalDetails.position = politician.professionalDetails.position?:line."position"?.trim()
+        politician.professionalDetails.institution = politician.professionalDetails.institution?:line."institution"?.trim()
+        politician.professionalDetails.constituency = politician.professionalDetails.constituency?:findConstituency(line)
+        politician.professionalDetails.region = politician.professionalDetails.region?:findRegion(line)
 
     }
 
@@ -367,11 +367,11 @@ class PoliticianService {
         if (!politician.careerDetails){
             politician.careerDetails = new CareerDetails()
         }
-        politician.careerDetails.profession = line."profession"
-        politician.careerDetails.cvLink =line."cvLink"
-        politician.careerDetails.declarationLink =line."declarationLink"
-        politician.careerDetails.university = line."university"
-        politician.careerDetails.school = line."school"
+        politician.careerDetails.profession = politician.careerDetails.profession?:line."profession"
+        politician.careerDetails.cvLink =politician.careerDetails.cvLink?:line."cvLink"
+        politician.careerDetails.declarationLink = politician.careerDetails.declarationLink?:line."declarationLink"
+        politician.careerDetails.university = politician.careerDetails.university?:line."university"
+        politician.careerDetails.school = politician.careerDetails.school?:line."school"
     }
 
     private Region findRegion(def line){
@@ -380,7 +380,7 @@ class PoliticianService {
     }
 
     private Region findConstituency(def line){
-        String constituencyCode = line."constituency"
+        String constituencyCode = line."constituency"?.trim()
         Region constituency
         if (constituencyCode){
             constituency = Region.findByIso3166_2(constituencyCode)
